requirements.txt
python-vlc>=3.0.21203
Flask>=3.0.2
pywebview>=5.1
